Event-store

+++ Source/Kernel/Events/EventLogSequenceNumber.cs
@@ -9,7 +9,7 @@ namespace Cratis.Events
     /// Represents the sequence number within an event log for an event.
     /// </summary>
     /// <param name="Value">The sequence number</param>
-    public record EventLogSequenceNumber(uint Value): ConceptAs<uint>(Value)
+    public record EventLogSequenceNumber(ulong Value): ConceptAs<ulong>(Value)
     {
         /// <summary>
         /// Gets the first sequence number.
@@ -17,10 +17,10 @@ namespace Cratis.Events
         public static readonly EventLogSequenceNumber First = 0u;
 
         /// <summary>
-        /// Implicitly convert from <see cref="uint"/> to <see cref="EventLogSequenceNumber"/>.
+        /// Implicitly convert from <see cref="ulong"/> to <see cref="EventLogSequenceNumber"/>.
         /// </summary>
         /// <param name="Value">Value to convert from.</param>
         /// <returns>A converted <see cref="EventLogSequenceNumber"/>.</returns>;
-        public static implicit operator EventLogSequenceNumber(uint Value) => new (Value);
+        public static implicit operator EventLogSequenceNumber(ulong Value) => new (Value);
     }
 }
